update user and video

+++ src/controllers/videoController.js
@@ -1,33 +1,45 @@
 import videoModel from "../models/video";
-
+import userModel from "../models/user";
 export const home = async (req, res) => {
   const videos = await videoModel.find({}).sort({ createdAt: "asc" });
-  console.log(videos);
   return res.render("home", { pageTitle: "Home", videos });
 };
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await videoModel.findById(id);
+  const video = await videoModel.findById(id).populate("owner");
+  console.log(video);
   if (!video) {
     return res.render("404", { pageTitle: "Video not found" });
   }
   return res.render("watch", { pageTitle: video.title, video });
 };
 export const getEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const video = await videoModel.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const video = await videoModel.exists({ _id: id });
+  const video = await videoModel.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await videoModel.findByIdAndUpdate(id, {
     title,
@@ -40,13 +52,22 @@ export const getUpload = (req, res) => {
   return res.render("upload", { pageTitle: "Upload Video" });
 };
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: videoUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await videoModel.create({
+    const newVideo = await videoModel.create({
       title,
       description,
+      videoUrl,
       hashtags: videoModel.formatHashtags(hashtags),
+      owner: _id,
     });
+    const user = await userModel.findById(_id);
+    user.videos.push(newVideo._id);
+    await user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -56,8 +77,16 @@ export const postUpload = async (req, res) => {
   }
 };
 export const deleteVideo = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
+  const video = await videoModel.findById(id);
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await videoModel.findByIdAndDelete(id);
+  await userModel.findByIdAndUpdate(_id, { $pull: { videos: id } });
   return res.redirect("/");
 };
